Make mapping numeric Python types to C++ types more robust
For some odd resons, for certain combinations of Python
versions/distribution and platforms, calling something like

```python3
f = hictkpy.File(...)
with pytest.raises(TypeError):
    f.fetch(count_type=str)
```

could lead to UnicodeDecodeError exceptions with the previous logic.
This likely occurred when falling back to parsing the string representation
of a type when inferring the C++ type from the Python type.